requirements.txt
bs4
discord.py